test: Seller, Money 테스트코드 추가
- 클래스 분리에 따른 테스트 코드 추가
