+++ src/test/java/LottosTest.java
@@ -22,15 +22,6 @@ class LottosTest {
     }
 
     @Test
-    void 천원이하의_금액을_입력하면_오류를_리턴한다() {
-        int money = 100;
-        assertThatThrownBy(() -> {
-            new Lottos(new Money(money), new MockStrategy(Arrays.asList(1, 2, 3, 4, 5, 6)));
-        }).isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("1000원 이상의 금액을 입력해주세요.");
-    }
-
-    @Test
     void 로또숫자의범위는_1에서_45여야한다() {
         int money = 10000;
         List<Integer> arrangeNumber = new ArrayList<>();
